+++ file_properties.txt
@@ -15,8 +15,8 @@
 #/uboone/data/users/gardiner/ntuples-stv/stv-data_extbnb_mcc9.1_v08_00_00_25_reco2_F_G_all_reco2.root 3 extBNB 214555174 0
 #
 # Full beam-on data
-/exp/annie/data/users/jminock/standard_tank_ntuples/PhaseIITree_large_unweighted_ntuple.root 1 onBNB 1895996 1.463e21
-#/exp/annie/data/users/jminock/standard_tank_ntuples/PhaseIITree_20k_ntuple.root 2 onBNB 20000 1.56e19
+#/exp/annie/data/users/jminock/standard_tank_ntuples/PhaseIITree_large_unweighted_ntuple.root 1 onBNB 1895996 1.463e21
+/pnfs/annie/persistent/users/jminock/v1_3_0_ntuples/PhaseIITree_400k_ntuple.root 1 onBNB 384680 1.5297e20
 #/exp/annie/data/users/jminock/standard_tank_ntuples/PhaseIITree_20k_ntuple.root 3 onBNB 20000 1.56e19
 #/uboone/data/users/gardiner/ntuples-stv/stv-run1_neutrinoselection_filt_numu_ALL.root 1 onBNB 36139233 1.62e20
 #/uboone/data/users/gardiner/ntuples-stv/stv-run2_neutrinoselection_filt_numu_ALL.root 2 onBNB 62045760 2.62e20
@@ -39,9 +39,9 @@
 #/uboone/data/users/gardiner/ntuples-stv/stv-prodgenie_bnb_intrinsice_nue_uboone_overlay_mcc9.1_v08_00_00_26_run3_reco2_reco2.root 3 nueMC
 #
 # numu MC samples
-/exp/annie/data/users/jminock/standard_tank_ntuples/PhaseIITree_large_unweighted_ntuple.root 1 numuMC
-#/exp/annie/data/users/jminock/standard_tank_ntuples/PhaseIITree_20k_ntuple.root 2 numuMC 20000 1.56e19
-#/exp/annie/data/users/jminock/standard_tank_ntuples/PhaseIITree_20k_ntuple.root 3 numuMC 20000 1.56e19
+#/exp/annie/data/users/jminock/standard_tank_ntuples/PhaseIITree_large_unweighted_ntuple.root 1 numuMC
+/pnfs/annie/persistent/users/jminock/v1_3_0_ntuples/PhaseIITree_400k_ntuple.root 1 numuMC
+#/exp/annie/data/users/jminock/standard_tank_ntuples/PhaseIITree_20k_ntuple.root 3 numuMC
 #/uboone/data/users/gardiner/ntuples-stv/stv-prodgenie_bnb_nu_uboone_overlay_mcc9.1_v08_00_00_26_filter_run1_reco2_reco2.root 1 numuMC
 #/uboone/data/users/gardiner/ntuples-stv/stv-prodgenie_bnb_nu_uboone_overlay_mcc9.1_v08_00_00_26_filter_run2_reco2_D1D2_reco2.root 2 numuMC
 #/uboone/data/users/gardiner/ntuples-stv/stv-prodgenie_bnb_nu_uboone_overlay_mcc9.1_v08_00_00_26_filter_run3_reco2_G_reco2.root 3 numuMC
