+++ systcalc.conf
@@ -1,3 +1,39 @@
+flux_expskin FluxRW weight_expskin_FluxUnisim 1
+flux_horncurrent FluxRW weight_horncurrent_FluxUnisim 1
+flux_kminus FluxRW weight_kminus_PrimaryHadronNormalization 1
+flux_kplus FluxRW weight_kplus_PrimaryHadronFeynmanScaling 1
+flux_kzero FluxRW weight_kzero_PrimaryHadronSanfordWang 1
+flux_nucleoninexsec FluxRW weight_nucleoninexsec_FluxUnisim 1
+flux_nucleonqexsec FluxRW weight_nucleonqexsec_FluxUnisim 1
+flux_nucleontotxsec FluxRW weight_nucleontotxsec_FluxUnisim 1
+flux_piminus FluxRW weight_piminus_PrimaryHadronSWCentralSplineVariation 1
+flux_pioninexsec FluxRW weight_pioninexsec_FluxUnisim 1
+flux_pionqexsec FluxRW weight_pionqexsec_FluxUnisim 1
+flux_piontotxsec FluxRW weight_piontotxsec_FluxUnisim 1
+flux_piplus FluxRW weight_piplus_PrimaryHadronSWCentralSplineVariation 1
+
+xsec_multi RW weight_All_UBGenie 1
+
+xsec_AxFFCCQEshape RW weight_AxFFCCQEshape_UBGenie 1
+xsec_DecayAngMEC RW weight_DecayAngMEC_UBGenie 1
+xsec_NormCCCOH RW weight_NormCCCOH_UBGenie 1
+xsec_NormNCCOH RW weight_Norm_NCCOH_UBGenie 1
+xsec_RPA_CCQE RW weight_RPA_CCQE_UBGenie 1
+xsec_ThetaDelta2NRad RW weight_ThetaDelta2NRad_UBGenie 1
+xsec_Theta_Delta2Npi RW weight_Theta_Delta2Npi_UBGenie 1
+xsec_VecFFCCQEshape RW weight_VecFFCCQEshape_UBGenie 1
+xsec_XSecShape_CCMEC RW weight_XSecShape_CCMEC_UBGenie 1
+
 MCstats MCstat
 
-total sum 1 MCstats
+xsec_unisim sum 9 xsec_AxFFCCQEshape xsec_DecayAngMEC xsec_NormCCCOH
+  xsec_NormNCCOH xsec_RPA_CCQE xsec_ThetaDelta2NRad xsec_Theta_Delta2Npi
+  xsec_VecFFCCQEshape xsec_XSecShape_CCMEC
+
+xsec_total sum 2 xsec_multi xsec_unisim
+
+flux_total sum 13 flux_expskin flux_horncurrent flux_kminus flux_kplus
+  flux_kzero flux_nucleoninexsec flux_nucleonqexsec flux_nucleontotxsec
+  flux_piminus flux_pioninexsec flux_pionqexsec flux_piontotxsec flux_piplus
+
+total sum 3 xsec_total flux_total MCstats
